requirements.txt
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
diffusers==0.25.0
transformers==4.35.0
accelerate==0.24.0
safetensors==0.4.0
huggingface_hub==0.19.0
xformers==0.0.22
opencv-python==4.8.1.78
pillow==10.0.0
numpy==1.24.0
runpod==1.6.0